requirements.txt
flask==2.0.1
werkzeug==2.0.1
langchain==0.0.184
langchain-community==0.0.9
langchain-text-splitters==0.0.1
langchain-huggingface==0.0.1
langchain-chroma==0.0.1
langchain-groq==0.0.1
python-dotenv==0.19.0
PyMuPDF==1.19.0

